Renamed application_directory to application_structure - more fitting.
